Разобьет отрисовку интерфейса ввода рейтинга на отдельные компоненты
