+++ project/src/const.js
@@ -4,3 +4,13 @@ export const AppRoute ={
   FAVORITES: '/favorites',
   ROOM: '/offer/:id',
 };
+
+export const ratings = [5, 4, 3, 2, 1];
+
+export const RatingSystem = {
+  1: 'terribly',
+  2: 'badly',
+  3: 'not bad',
+  4: 'good',
+  5: 'perfect',
+};
